CCB-245-Airborne
The request is to deprecate the permissible value "Airborne" from the  	enumerated lists for the attributes <Observing_System_Component.type> and  	<Product_Context.product_data_object> and to deprecate the <TNDO_Context>  	subclass <Airborne>.

The permissible values “Aircraft”, “Balloon”, and “Suborbital Rocket” are to be made values of the enumerated list for the attribute <Instrument_Host.type>. These values were deprecated in Observing_System_Component.

This change makes the use of “Aircraft”, “Balloon”, and “Suborbital rocket”  	  consistent with the way "Spacecraft", "Rover", and "Lander" are used for mission-based host products.

Also required deprecating "is_airborne" in <Internal_Reference.reference_type>