CCB-242-Field_Campaign
Add 'Field Campaign' to Investigation Type to accommodate planetary analog field studies.

Add 'Field Campaign' definition:
	A set of experiments and/or observations with a clearly defined purpose that is coordinated among one or more individuals and/or one or more organizations. The campaign is characterized primarily by its in situ terrestrial field work.

Investigation definitions (IM) should be amended to the following:

Investigation - A set of experiments and/or observations with a clearly defined purpose.

Individual - A set of experiments and/or observations with a clearly defined purpose that is under the direction of a single individual. For example, a series of related laboratory experiments/observations conducted by a small university team and funded through a single proposal could be an Individual Investigation. Library research by one person on previously published results to distill common conclusions could also be an Individual Investigation.

Mission - A set of experiments and/or observations with a clearly defined purpose that is under the direction of an organization (such as a national space agency) and uses one or more vehicles of similar design. A mission is usually conducted by a national or international space agency using a small number (often one) of spacecraft especially designed and built for the purpose. For example: (1) the NASA Voyager mission launched two spacecraft to Jupiter and Saturn; (2) the Deep Impact (DI) mission launched a spacecraft, part of which impacted the comet Tempel 2; and (3) the EPOXI mission redefined the objectives of the surviving DI hardware for further exploration of the solar system.

Observing Campaign - A set of experiments and/or observations with a clearly defined purpose that is coordinated among one or more individuals and/or one or more organizations. The campaign is characterized primarily by its remote sensing observations. For example, the International Halley Watch mobilized observatories around the world for collection of data during the most recent apparition of Comet Halley. The Shoemaker-Levy-9 (SL9) campaign used both Earth-based and spacecraft instruments to monitor the SL9 impact on Jupiter.

		Field Campaign - A set of experiments and/or observations with a clearly defined purpose that is coordinated among one or more individuals and/or one or more organizations. The campaign is characterized primarily by its in situ terrestrial field work. Field campaigns may include multiple types of in situ measurements and/or remote sensing for a unified scientific purpose. An example would be for the characterization of a dust devil environment using in situ pressure/temperature sensors and anemometers, remote sensing imagers/cameras, mounted sediment traps, etc., at a single/multiple field site(s).

		Other - Any other set of experiments and/or observations with a unifying theme.