CCB227-Source_Product_External (#20)
1. Add a Schematron rule with the following logic: if DOI is not present,  curating facility must be present. Else, if curating facility is not present, DOI must be present.

2. Update the Schema to use this new definition for the Source_Product_External class:

The Source_Product_External class is used to reference one or more source products (a product containing input data for the creation of this product) outside the PDS4 Registry that have a common reference_type, doi, curating facility, and/or description. At least one of doi or curating facility must be provided. All source products listed within a single Source_Product_External class must correspond to the same doi and/or curating facility.