190301_CCBnnn-InconsistentLineEndings_CRLF_Schema (#31)
In the WrapText routine, <CR><LF> are being written explicitly for each line ending in the text block on both Windows and Mac machines.

Solution is to partition the text block into an array of lines and use printf to write each line. The printf untility writes the proper line ending.