PDS F2F Action Items - PDSF2F-2 - LDDTool: Add details to the version (-v) option (#51)
PDS F2F Action Items - PDSF2F-2 - LDDTool: Add details to the version (-v) option so we know what is the differences between versions
		LDD Versioning – there are three version identifiers:
		IM version ID,    ldd_version_id,    product version_id
Partial Solution: The -v option currently displays the version of the software, statically defined in the code and then exits. The ldd_version_id and product_version_id are only know after processing the IngesteLDD file. The short term fix was to add this information to both the LDDTool Report and the LDDTool log (system.out messages)