LDDTool does not handle "unbounded" maximum cardinality in Choice blocks correctly. (#181)
LDDTool does not handle "unbounded" maximum cardinality in Choice blocks correctly.

When defining a Choice block in Ingest_LDD, the character "*" is used as a value to represent an unlimited upper bound for cardinality. This value needs to be converted to "unbounded" when writing to an XML SChema file. This error only occurred in the case of the Choice block.

Resolves #180