LDDTool: erroneously replaces TBD values with null in JSON output
LDDTool: erroneously replaces TBD values with null in JSON output. When running LDDTool to produce JSON output, <definition> values in <DD_Attribute> that start with TBD are replaced with null in the output.

Since LDDTool uses "TBD" as an equivalent for "null" to indicate that a user value was not provided, the solution is to prepend a user supplied "TBD" with "_" resulting in "_TBD".

Resolves #227