CCB-311 - Value meanings for <rule_type> permissible values should be rewritten
Modify the value meanings of the permissible values of <rule_type> to be more meaningful.

Assert - Assert rules are usually used for error detection. If the test fails, the associated message is displayed.

Assert Every - An Assert Every rule is an extension to the Assert rule that uses the "every" qualifier to indicate every member of a defined set.

Assert If - An Assert IF rule is an extension to the Assert rule that limits the rule to the "if-then" pattern.

Report - Report rules are usually used for information-gathering. If the test succeeds, the associated message is displayed.

Resolves #296
Refs CCB-311