CCB-300 - Apparently deprecated units of measure are not actually deprecated (#307)
Apparently deprecated units of measure are not actually deprecated.

Units of measure that are indicated as being deprecated are not actually deprecated. Add schematron rules to test for and flag an error for deprecated units of measure in a label attribute. These schematron rules have been already created and can be reviewed in the available download.

Resolves #255
Refs CCB-300