CCB-315 - "PDS3" is an allowed parsing standard for Bundle documentation file
Remove "PDS3" from Stream_Text as allowable string in Product_Bundle.

For this ticket, this would be only pushing a rule for Product_Bundle to NOT allow "PDS3".

The request is to add a Schematron rule that would disallow "PDS3" to be used as a standard value for the attribute <Stream_Text.parsing_standard_id> in the context of <Product_Bundle.File_Area_Text.Stream_Text>. When the bundle’s <File_Area_Text> is used to describe a README file the standard value for the attribute <Stream_Text.parsing_standard_id> should be chosen from the set {“7-Bit ASCII Text”,”UTF-8 Text”}.

Resolves #250
Refs CCB-315