Issue with DD_Associate_External_Class
Issue with DD_Associate_External_Class when trying to reference pds:Internal_Reference and pds:Local_Internal_Reference

The issue reported seems to be associated with an earlier release of LDDTool. The reported problem could not be duplicated. However on inspection it was discovered that the Assert message was formatted wrong. The message formatting was corrected.

Resolves #391
Refs CCB-256