Issue-#398 - Resolve references across a partitioned namespace - 2
LDDTool was not resolving references across the separate files, even through both files had the same namespace. This issue is a result of the Geometry LDD partitioning task converting the single geom IngestLDD file into two or more IngestLDD files to reduce complexity. The issue was resolved by creating a single global intermediate storage for the defined classes and attributes.

Error message was

 >>> ERROR Class:0001_NASA_PDS_1.cart.Map_Projection_Lander Association:geom.Coordinate_Space_Reference Class:geom.Coordinate_Space_Reference - Missing Component Class
 >>> ERROR Association: geom.Coordinate_Space_Reference - Missing Component - Reference Type: component_of

Resolves #398
Refs CCB-256