CCB-329 - Broaden Definition of Attribute aperture (#405)
CCB-329 - The aperture attribute is defined for optical instruments with circular light collecting openings. Telescopes at radio wavelengths are not necessarily circular and may be arrays. Broadening the definition will make the attribute more useful at all wavelengths.

Change definition of aperture to: "The aperture attribute provides a measure of the effective collecting area of the telescope -- its diameter (if single and circular) or its equivalent diameter (if non-circular and/or an array).  For purposes of this definition, aperture efficiency is assumed to be 100 percent."

Resolves #403
Refs CCB-329

Co-authored-by: John Hughes <John.S.Hughes@jpl.nasa.gov>