#416 - LDDTool: New namespace id has been specified:ml
LDDTool reports a WARNING stating that a new namespace id has been specified, when in fact the namespace had been added to the config.properties file.

Obsolete code was causing the warning to be written. The obsolete code has been removed. Note that the warning was occurring for all LDD namespaces. However the warning was typically not noticed since most LDD runs have numerous warnings.

Resolves #416