Issue #428 - IMTool Refactoring - Phase 2
Issue #428 - IMTool Refactoring - Phase 2 - Move deprecated values to Protege ontology file.

This is phase 2 of Issue 306 - Refactor Class information from config to Data Dictionary - This phase focuses on moving deprecation information from the code to the Protege ontology. Also the version identifier for classes and attributes are being moved to the ontology file.

Resolves #428