Issue #269 - IM Spec Inconsistent Ordering Permissible Values
Issue #269 - The PDS4 IM Specification Document is not consistent in the ordering of Permissible Values.

The use of alphabetical order within the IM specification is wide spread. Sometimes, for no obvious reason the convention is dropped which can lead the unwary (read lazy) to oversights.

See Extended Values for: pds:Product_Context/pds:Reference_List/pds:Internal_Reference:
resource_to_target - The resource is associated to a target
target_to_document - The target is associated to a document
target_to_instrument - The target is associated to an instrument
target_to_instrument_host - The target is associated to an instrument_host

Resolves #269