CCB-343 - Revise Product_Metadata_Supplemental Desc
CCB-343 - Revise Product_Metadata_Supplemental  Description

The request is to remove the restriction that Product_Metadata_Supplemental may only be associated with one collection. Removing this restriction requires changing text in the description of the class in the IM and associated text in the Standards Reference. This change will improve the effective use of Product_Metadata_Supplemental.

Deprecated code was also removed from XML4LabelSchemaDOM

Resolves #448
Refs CCB-343