Issue #468 : LDDTool is generating multiple Schematron rules to test the same condition
Issue #468 : LDDTool is generating multiple Schematron rules to test the same condition

The IngestLDD for the DISP LDD includes DD_Permissible_Value  for the values “true” and “false”, presumably to provide value meanings for the values of the attributes Movie_Display_Settings/disp:loop_flag and disp:Movie_Display_Settings/disp:loop_back_and_forth_flag. Since these attributes each have a set of enumerated values, LDDTool assume that rules need to be written to validate the enumerated values. However, since the value type of each attribute is ASCII_Boolean, the values of the attributes are validated using the definition of ASCII_Boolean in the common dictionary. There is no need for both types of validation so a test was added to preclude the rules for attributes of type ASCII_Boolean  that also have enumerated values.

Resolves #468