CCB209 - Fix the PDS4 IM class hierarchy
CCB209 - Fix the PDS4 IM class hierarchy

The PDS4 Information Model (IM) has been modified to conform to the Open Archival Information System (OAIS) Reference Model. In particular each entity defined in the IM is now defined as an Information Object, a composite of metadata and a data object. More specifically, each IM entity that describes data is now associated with the defined class, "Digital_Object". This change does not effect the files that are generated for operations, for example the .xsd and .sch files.

Resolves #159
Refs CCB-209