Issue #507 : Bug generating BepiColombo SERENA dictionary
Issue #507 : Bug generating BepiColombo SERENA dictionary

LDDTool creates namespaces using the namespace id as defined in the Ingest_LDD file. However for PSA namespaces, the namespace id needs to be converted from a single string, to a list of directories. For example the namespace id "bc_mpo_srn" must be replaced with "bc/mpo/srn" for use in the namespace.

Resolves #507