Issue #506 : hyb2 has incorrect namespace for IM v1.14  - Redo (#511)
Issue #506 : hyb2 has incorrect namespace for IM v1.14  - Redo

Synchronized all prior versions of the config.properties file to the latest set of namespace declarations.

Prior to this change, a new namespace was only added to the latest version of the config.properties files. However the namespace "hyb2" was required while running LDDTool using a prior version of the IM, v1.14.0.0 (1E00). The first fix was to simply update config.properties v1.14.0.0. However this did not work since the structure of the config.properties file had changed.

Further analysis suggested that all prior versions of the config.properties files should have the same set of namespace definitions. The impact of this more comprehensive change is that the LDDTool output listing will simply include a larger list of available namespaces for LDDTool runs with prior versions of the IM.

Resolves #506

Co-authored-by: John Hughes <jsh416@gmail.com>