Issue #552 : add new namespace "sbn" (#553)
Issue #552 : add new namespace "sbn"

The namespace "sbn" was added to the registry on 2012-04-03. Changes resulting from this request are minor changes to the name and description of the namespace. There were no changes necessary for the config.properties file.

This dictionary will provide classes to support the documentation, support, discovery, and reuse of data from, by, and for small bodies research.

@c-suh  [namespace-registry] add new namespace "sbn" #552

Resolves #552

Co-authored-by: John Hughes <jsh416@gmail.com>