Issue #563 : [namespace-registry] add new namespace "sb"
Issue #563 : [namespace-registry] add new namespace "sb"

Add the namespace "sb" to the Namespace Registry and the config.properties file.

This dictionary will provide classes to support the documentation, support, discovery, and reuse of data from, by, and for small bodies research.

This is a redo of 552. Instead of "sbn" the namespace id should be "sb".

@c-suh  [namespace-registry] add new namespace "sb" #563

Resolves #563